backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.4
openai==1.77.0
PyMuPDF==1.25.5
python-docx==1.1.2
python-multipart==0.0.20
python-dotenv==1.1.0
